pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.21.0",
    "aiohttp>=3.12.15",
    "python-dotenv>=1.0.1",
    "requests>=2.32.5",
    "telethon>=1.40.0",
]
